refactor: created a common component for layer type switch
